requirements.txt
Django==5.0.7
django-allauth==0.61.1
langchain==0.3.27
langchain_google_genai==2.1.8
torch==2.7.1
langchain_community==0.3.27
getenv==0.2.0
PyJWT=2.10.1